Former-commit-id: f87d5f6dfe5068a66e048956d922f167cabe5df0

+++ Sources/Internal/Render/2D/TextBlockDistanceRender.cpp
@@ -88,7 +88,7 @@ void TextBlockDistanceRender::Prepare(Texture *texture /*= NULL*/)
     uint32 charCount = 0;
     if (!textBlock->isMultilineEnabled || textBlock->treatMultilineAsSingleLine)
     {
-        charCount = textBlock->visualText.length();
+        charCount = static_cast<uint32>(textBlock->visualText.length());
     }
     else
     {
